requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
pytest
loguru